Consolidate gltools utils
