--- modules/gltools/src/polyfill/assert.js
@@ -1,5 +1,0 @@
-export default function assert(condition, message) {
-  if (!condition) {
-    throw new Error(message || 'gltools: assertion failed.');
-  }
-}
