--- modules/gltools/src/state-tracker/utils/assert.js
@@ -1,6 +1,0 @@
-// Avoid bundling assert polyfill module
-export default function assert(condition, message) {
-  if (!condition) {
-    throw new Error(message || 'luma.gl: assertion failed.');
-  }
-}
