--- modules/gltools/src/state-tracker/utils/deep-array-equal.js
@@ -1,16 +1,0 @@
-export default function deepArrayEqual(x, y) {
-  if (x === y) {
-    return true;
-  }
-  const isArrayX = Array.isArray(x) || ArrayBuffer.isView(x);
-  const isArrayY = Array.isArray(y) || ArrayBuffer.isView(y);
-  if (isArrayX && isArrayY && x.length === y.length) {
-    for (let i = 0; i < x.length; ++i) {
-      if (x[i] !== y[i]) {
-        return false;
-      }
-    }
-    return true;
-  }
-  return false;
-}
