--- modules/gltools/src/state-tracker/utils/index.js
@@ -1,3 +1,0 @@
-export {default as assert} from './assert';
-export {default as deepArrayEqual} from './deep-array-equal';
-export {isWebGL2, isObjectEmpty} from './utils';
