+++ modules/gltools/test/polyfill/polyfills/get-parameter-polyfill.spec.js
@@ -1,4 +1,4 @@
-import {getParameterPolyfill} from '@luma.gl/gltools/polyfill/polyfills/get-parameter-polyfill';
+import {getParameterPolyfill} from '@luma.gl/gltools/polyfill/get-parameter-polyfill';
 import GL from '@luma.gl/constants';
 import test from 'tape-catch';
